Code - Web Application/cloud_functions/metadata_extractor/requirements.txt
pillow==10.3.0
pandas==2.2.1
google-cloud-storage==2.16.0
google-cloud-pubsub==2.10.0